🚧 14X73Y task: apply external agent result
