+++ .agentplane/tasks/202609121655-14X73Y/pr/diffstat.txt
@@ -1,3 +1,4 @@
- .../src/commands/task/scope-extend.test.ts         | 110 +++++++++++++++++++++
- .../agentplane/src/commands/task/scope-extend.ts   |  37 ++++++-
- 2 files changed, 145 insertions(+), 2 deletions(-)
+ .../src/commands/task/scope-extend.command.test.ts |  95 +++++++++++
+ .../src/commands/task/scope-extend.test.ts         | 188 ++++++++++-----------
+ .../agentplane/src/commands/task/scope-extend.ts   |  37 +++-
+ 3 files changed, 224 insertions(+), 96 deletions(-)
