+++ .agentplane/tasks/202609121655-14X73Y/pr/review.md
@@ -6,14 +6,14 @@ Created: 2026-09-12T16:56:50.829Z
 
 - Task: `202609121655-14X73Y`
 - Title: Fix exact WorkItem-only scope extension when the global contract is already satisfied
-- Status: DONE
+- Status: DOING
 - Branch: `task/202609121655-14X73Y/fix-exact-workitem-only-scope-extension-when-the`
 - Canonical task record: `.agentplane/tasks/202609121655-14X73Y/README.md`
 
 ## Verification
 
-- State: ok
-- Note: Verified: CLI-owned checks passed before independent EVALUATOR review.
+- State: pending
+- Note: Invalidated by USER-approved execution scope extension.
 - Canonical workflow state lives in the task README.
 
 ## Handoff Notes
@@ -29,9 +29,10 @@ Created: 2026-09-12T16:56:50.829Z
 - Head: computed live by `agentplane pr check` / `agentplane integrate`
 
 ```text
- .../src/commands/task/scope-extend.test.ts         | 110 +++++++++++++++++++++
- .../agentplane/src/commands/task/scope-extend.ts   |  37 ++++++-
- 2 files changed, 145 insertions(+), 2 deletions(-)
+ .../src/commands/task/scope-extend.command.test.ts |  95 +++++++++++
+ .../src/commands/task/scope-extend.test.ts         | 188 ++++++++++-----------
+ .../agentplane/src/commands/task/scope-extend.ts   |  37 +++-
+ 3 files changed, 224 insertions(+), 96 deletions(-)
 ```
 
 </details>
